+++ source/set.js
@@ -1,17 +1,13 @@
 import { always } from './always'
 import { over } from './over'
+import {curry} from './curry'
 
-export function set(
+function setFn(
   lens, replacer, x
 ){
-  if (arguments.length === 1) return (_v, _x) => set(
-    lens, _v, _x
-  )
-  if (arguments.length === 2) return _x => set(
-    lens, replacer, _x
-  )
-
   return over(
     lens, always(replacer), x
   )
 }
+
+export const set = curry(setFn)